@@ -3,16 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@google-pay/button-react": "^3.0.0",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.8.7",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.19.2",
     "bootstrap": "^4.5.1",
-    "react": "^16.13.1",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-apple-pay-button": "0.0.1-alpha.0",
     "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-paypal-button-v2": "^2.6.2",
-    "react-router-dom": "^5.2.0",
+    "react-dom": "^18.2.0",
+    "react-paypal-button-v2": "^2.6.3",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "3.4.1"
   },
   "scripts": {
