@@ -1,123 +1,162 @@
-import React, { Component } from "react";
+import React, { useEffect, useState } from "react";
 import "./App.css";
 
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import {
+  BrowserRouter as Router,
+  Route,
+  Routes,
+  useLocation,
+} from "react-router-dom";
 
-import Header from "./components/Header";
-import Footer from "./components/Footer";
-import ItemList from "./components/ItemList";
+import Header from "./components/Header/Header";
+import Footer from "./components/Footer/Footer";
+import ItemList from "./components/ItemList/ItemList";
 import Item from "./components/Item";
-import Cart from "./components/Cart";
-import Signin from "./components/Signin";
-import Join from "./components/Join";
-import Checkout from "./components/Checkout";
+import Cart from "./components/Cart/Cart";
+import Signin from "./components/Account/Signin";
+import Join from "./components/Account/Join";
+import Checkout from "./components/Checkout/Checkout";
 import Profile from "./components/Profile";
-import OrderHistory from "./components/OrderHistory";
+import OrderHistory from "./components/OrderHistory/OrderHistory";
+import Favourite from "./components/Favourite/Favourite";
 
-import { API_BASE_URL } from "./constants";
+import {
+  getAllItemsInCart,
+  getCartId,
+  getCurrentUser,
+} from "./components/Util/Requests";
 
-import axios from "axios";
+export default function App() {
+  const [currentUser, setCurrentUser] = useState({});
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const [customerId, setCustomerId] = useState(
+    localStorage.getItem("CUSTOMER_ID")
+  );
+  const [cartId, setCartId] = useState(localStorage.getItem("CART_ID"));
+  const [itemsInCart, setItemsInCart] = useState([]);
 
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      currentUser: {},
-      isAuthenticated: false,
-      customerId: localStorage.getItem("CUSTOMER_ID"),
-      cartId: localStorage.getItem("CART_ID"),
-    };
-  }
+  const location = useLocation();
 
-  componentDidMount() {
-    console.log("cart id: ", localStorage.getItem("CART_ID"));
-    console.log("customer id: ", localStorage.getItem("CUSTOMER_ID"));
-    console.log("token: ", localStorage.getItem("ACCESS_TOKEN"));
-    this.loadUser();
-    this.assignCart();
-  }
+  console.log("pathname", location.pathname);
+  console.log("search", location.search);
 
-  assignCart = () => {
-    if (this.state.cartId === null) {
-      axios
-        .get(`${API_BASE_URL}/cart/getId`)
-        .then((response) => response.data)
-        .then((data) => {
-          this.setState({ cartId: Math.max(...data) + 1 });
-          localStorage.setItem("CART_ID", Math.max(...data) + 1);
-        });
+  const assignCart = async () => {
+    if (!cartId) {
+      const data = await getCartId();
+      setCartId(data);
+      localStorage.setItem("CART_ID", data);
     }
   };
 
-  loadUser = () => {
-    axios
-      .get(`${API_BASE_URL}/customer/${this.state.customerId}`, {
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem("ACCESS_TOKEN")}`,
-        },
-      })
-      .then((response) => response.data)
-      .then((data) => {
-        this.setState({ currentUser: data, isAuthenticated: true });
-      })
-      .catch((error) => console.log("Not logged in"));
+  const findAllItemsInCart = async () => {
+    const data = await getAllItemsInCart(cartId);
+    setItemsInCart(data);
   };
 
-  render() {
-    return (
-      <Router>
-        <Header
-          isAuthenticated={this.state.isAuthenticated}
-          currentUser={this.state.currentUser}
-        />
-
-        <Switch>
-          <Route exact path="/" exact component={ItemList} />
-          <Route
-            exact
-            path="/department/:department"
-            exact
-            component={ItemList}
-          />
-          <Route exact path="/search/:search" exact component={ItemList} />
-          <Route path="/item" exact component={Item} />
-          <Route path="/cart" exact component={Cart} />
-          <Route
-            path="/signin"
-            exact
-            component={this.state.isAuthenticated ? ItemList : Signin}
-          />
-          <Route
-            path="/join"
-            exact
-            component={this.state.isAuthenticated ? ItemList : Join}
-          />
-          <Route
-            path="/checkout"
-            render={(props) => (
-              <Checkout
-                isAuthenticated={this.state.isAuthenticated}
-                currentUser={this.state.currentUser}
-                {...props}
-              />
-            )}
-          />
-          <Route
-            path="/profile"
-            render={(props) => (
-              <Profile
-                isAuthenticated={this.state.isAuthenticated}
-                currentUser={this.state.currentUser}
-                {...props}
-              />
-            )}
-          />
-          <Route path="/history" exact component={OrderHistory} />
-        </Switch>
-        <Footer />
-      </Router>
+  const loadUser = async () => {
+    const data = await getCurrentUser(
+      customerId,
+      sessionStorage.getItem("ACCESS_TOKEN")
     );
-  }
-}
+    if (!data || !Object.keys(data).length) {
+      setIsAuthenticated(false);
+    } else {
+      setCurrentUser(data);
+      setIsAuthenticated(true);
+    }
+  };
 
-export default App;
+  useEffect(() => {
+    console.log("cart id: ", localStorage.getItem("CART_ID"));
+    console.log("customer id: ", localStorage.getItem("CUSTOMER_ID"));
+    console.log("token: ", sessionStorage.getItem("ACCESS_TOKEN"));
+    loadUser();
+    assignCart();
+    findAllItemsInCart();
+  }, []);
+
+  return (
+    <>
+      <Header
+        isAuthenticated={isAuthenticated}
+        currentUser={currentUser}
+        itemsInCart={itemsInCart}
+      />
+      <Routes>
+        <Route path="/">
+          <Route index element={<ItemList />} />
+          <Route path="browse" element={<ItemList />}>
+            <Route path=":department" element={<ItemList />} />
+          </Route>
+          <Route path="search" element={<ItemList />}>
+            <Route path=":search" element={<ItemList />} />
+          </Route>
+          <Route
+            path="item"
+            element={
+              <Item
+                cartId={cartId}
+                itemsInCart={itemsInCart}
+                setItemsInCart={setItemsInCart}
+              />
+            }
+          />
+
+          <Route
+            path="cart"
+            element={
+              <Cart
+                isAuthenticated={isAuthenticated}
+                itemsInCart={itemsInCart}
+                setItemsInCart={setItemsInCart}
+              />
+            }
+          />
+          <Route
+            path="checkout"
+            element={
+              <Checkout
+                isAuthenticated={isAuthenticated}
+                currentUser={currentUser}
+                itemsInCart={itemsInCart}
+                setItemsInCart={setItemsInCart}
+              />
+            }
+          />
+          <Route
+            path="signin"
+            element={
+              isAuthenticated ? <ItemList /> : <Signin cartId={cartId} />
+            }
+          />
+          <Route
+            path="join"
+            element={isAuthenticated ? <ItemList /> : <Join cartId={cartId} />}
+          />
+
+          <Route
+            path="profile"
+            element={
+              <Profile
+                isAuthenticated={isAuthenticated}
+                currentUser={currentUser}
+              />
+            }
+          />
+          <Route path="history" exact element={<OrderHistory />} />
+          <Route
+            path="favourite"
+            element={
+              <Favourite
+                cartId={cartId}
+                itemsInCart={itemsInCart}
+                setItemsInCart={setItemsInCart}
+              />
+            }
+          />
+        </Route>
+      </Routes>
+      <Footer />
+    </>
+  );
+}
