@@ -1,3 +1,3 @@
-// export const API_BASE_URL = "http://localhost:8080/api";
+export const API_BASE_URL = "http://localhost:8080/api";
 
-export const API_BASE_URL = "https://ecommerce-spring-react.herokuapp.com/api";
+// export const API_BASE_URL = "https://ecommerce-spring-react.herokuapp.com/api";
